Add Interactions features
Add interactions features to allow logging communications and in-person
interactions with invidividuals. This includes a new Interaction Channel,
a general Interaction Component, screens to view interactions from a person
profile, workflows to enter interactions for an individual or in bulk,
and custom grid actions on the group viewer and data view screens.

A custom UI control and Field Types were added create the desired user
experience in the workflows