Update recommended department structure, add miss data
This update deletes the department defined type
in favor of a Department group type. Each department in
the organization is a top-level group of type of Department.
Instead of the Area>Region>Team hierarchy being under the
Missionaries group of a Department, there is a new default
Teams group where the heirarchy should live.

Each department can have a Missionaries and Staff group that
each can be synced with a dataview that has that departments
missionaries and staff.

This also adds a Communication Risk Override attribute to teams.
This also turns on Group History for Teams, Missionaries, and Staff.

The ABWECampusGroupAndRolePicker is updated to disallow choosing
the base Missionaries group.