Add Trip Planning features
This update adds a new Picker control, DataView Filter, and Report field
for Trip Planning functionality in Rock. It also adds a Lava
template for use in the trip planning tool.

The Picker control uses the Google Places API to allow you to search
for and choose a set of longitude and latitude coordinates by
place name. The DataView filter of DistanceFromWithBusiness uses this
field to allow you to choose a place and select all person records
(including businesses) that have an address within a certain distance.
It is an adaptation of the Distance From filter and GeoPicker control.

The Report filter allows you to select the LastInteraction WITH an
individual, rather than an Interaction BY an individual.

The lava templates can be used in a DynamicReport block and allow you
to create a map from latitude and longitude columsn in the report. See
the template for more info. It uses a MutationObserver to watch for changes
to the dynamic report and repopulate the map based on the people in the
grid, so all the mapping happens client-side.